Feat:add check author, existing diary

+++ back/src/controllers/diaryController.ts
@@ -3,7 +3,7 @@ import {
   createDiaryService,
   deleteDiaryService,
   getAllDiaryService,
-  getDiaryByDiaryIdService,
+  getOneDiaryService,
   getDiaryByMonthService,
   getAllMyDiariesService,
   updateDiaryService,
@@ -11,6 +11,8 @@ import {
   mailService,
   selectedEmoji,
   searchDiaryService,
+  getDiaryByDateService,
+  checkAuthor,
 } from '../services/diaryService';
 import { IRequest } from 'types/user';
 import { plainToClass } from 'class-transformer';
@@ -30,12 +32,21 @@ export const createDiary = async (
   res: Response,
   next: NextFunction,
 ) => {
-  // const fileUrls = res.locals.myData;
-  const fileUrls = ['test'];
+  const fileUrls = res.locals.myData;
+  const {
+    body: inputData,
+    user: { id: userId },
+  } = req;
 
-  // const { emoji,...inputData } = req.body;
-  const inputData = req.body;
+  const { createdDate } = inputData;
 
+  // 해당 날짜에 다이어리 존재하는지 체크
+  const checkExistedDiary = await getDiaryByDateService(userId, createdDate);
+
+  if (checkExistedDiary) {
+    console.log('이미 존재하지롱');
+    return res.json('이미 존재하는 다이어리가 있어용 메롱 ');
+  }
   const diaryInput = plainToClass(DiaryValidateDTO, inputData);
 
   // TODO 밸리데이터 수정 필요
@@ -47,11 +58,6 @@ export const createDiary = async (
       return error.constraints;
     });
   }
-
-  //throw generateError(400, errors[0].constraints);
-  // console.log('!!!!!!!!!!!!', errors[0].constraints);
-
-  const { id: userId } = req.user;
 
   const createdDiary = await createDiaryService(userId, inputData, fileUrls);
   console.log(createDiary);
@@ -72,8 +78,8 @@ export const getAllMyDiaries = async (
 ) => {
   //authorId
   const { id: userId } = req.user;
-  const page = Number(req.query.page) || 1;
-  const limit = Number(req.query.limit) || 8;
+  const page = parseInt(req.query.page as string) || 1;
+  const limit = parseInt(req.query.limit as string) || 8;
 
   const myDiaries = await getAllMyDiariesService(userId, page, limit);
 
@@ -86,9 +92,10 @@ export const getDiaryByDate = async (
   next: NextFunction,
 ) => {
   //authorId
+
   const { userId } = req.params;
-  const year = Number(req.query.year);
-  const month = Number(req.query.month);
+  const year = parseInt(req.query.year as string);
+  const month = parseInt(req.query.month as string);
   const MonthlyDiary = await getDiaryByMonthService(userId, year, month);
 
   return res.status(MonthlyDiary.status).json(MonthlyDiary);
@@ -100,14 +107,17 @@ export const getDiaryByDate = async (
  * @param next
  * @returns
  */
-export const getDiaryByDiaryId = async (
+export const getOneDiary = async (
   req: IRequest,
   res: Response,
   next: NextFunction,
 ) => {
-  const { diaryId } = req.params;
-  const { id: userId } = req.user;
-  const diary = await getDiaryByDiaryIdService(userId, diaryId);
+  const {
+    params: { diaryId },
+    user: { id: userId },
+  } = req;
+
+  const diary = await getOneDiaryService(userId, diaryId);
 
   return res.status(diary.status).json(diary);
 };
@@ -124,11 +134,17 @@ export const getOtherUsersDiary = async (
   res: Response,
   next: NextFunction,
 ) => {
-  const { id: userId } = req.user;
-  const { select } = req.query; // friend or all
-  const page = Number(req.query.page) || 1;
-  const limit = Number(req.query.limit) || 8;
-  const { emotion } = req.query;
+  /**
+   * #swagger.tags = ['Diary']
+   * #swagger.summary = '친구 요청'
+   */
+  const {
+    query: { select, emotion },
+    user: { id: userId },
+  } = req;
+  const page = parseInt(req.query.page as string) || 1;
+  const limit = parseInt(req.query.limit as string) || 8;
+
   // diary 데이터 가져오기
 
   const otherUsersDiary =
@@ -143,10 +159,19 @@ export const updateDiary = async (
   res: Response,
   next: NextFunction,
 ) => {
-  const { id: userId } = req.user;
-  const { diaryId } = req.params;
-  const inputData = req.body;
+  const {
+    body: inputData,
+    params: { diaryId },
+    user: { id: userId },
+  } = req;
+  // const { id: userId } = req.user;
+  // const { diaryId } = req.params;
+  // const inputData = req.body;
   const { deleteData, ...updatedData } = inputData;
+  const isMatchedAuthor = await checkAuthor(diaryId, userId);
+
+  if (!isMatchedAuthor) return res.status(403).json('작성자가 아닙니다.');
+
   const diaryInput = plainToClass(DiaryValidateDTO, updatedData);
 
   // TODO 밸리데이터 수정 필요
@@ -170,8 +195,15 @@ export const deleteDiary = async (
   res: Response,
   next: NextFunction,
 ) => {
-  const { diaryId } = req.params;
-  const { id: userId } = req.user;
+  const {
+    params: { diaryId },
+    user: { id: userId },
+  } = req;
+
+  const isMatchedAuthor = await checkAuthor(diaryId, userId);
+
+  if (!isMatchedAuthor) return res.status(403).json('작성자가 아닙니다.');
+
   const deletedDiary = await deleteDiaryService(userId, diaryId);
 
   return res.status(deletedDiary.status).json(deletedDiary);
@@ -218,8 +250,10 @@ export const searchDiary = async (
   res: Response,
   next: NextFunction,
 ) => {
-  const search = req.body;
-  const searchDiary = await searchDiaryService(search);
+  const page = parseInt(req.query.page as string) || 1;
+  const limit = parseInt(req.query.limit as string) || 8;
+  const { search } = req.body;
+  const searchDiary = await searchDiaryService(search, page, limit);
 
   return res.status(200).json(searchDiary);
 };
